pyproject.toml
[project]
name = "myasr"
version = "0.1.0"
description = "语音转文字 + LLM文本处理 API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "anthropic>=0.30.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "mangum>=0.20.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[tool.uv]
dev-dependencies = []
